requirements.txt
astor==0.8.1
asttokens==2.4.1
attrs==23.1.0
autogluon.common==1.1.0b20240411
autogluon.core==1.1.0b20240411
autogluon.features==1.1.0b20240411
autogluon.tabular==1.1.0b20240411
bokeh==3.3.1
boto3==1.34.107
botocore==1.34.107
category-encoders==2.6.3
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.5
debugpy==1.8.1
decorator==5.1.1
descartes==1.1.0
distlib==0.3.7
efficient-apriori==2.0.3
et-xmlfile==1.1.0
executing==2.0.1
factor-analyzer==0.5.0
fastjsonschema==2.19.1
filelock==3.13.1
fiona==1.9.5
fonttools==4.45.0
formulaic==0.6.6
geopandas==0.14.1
identify==2.5.32
idna==3.7
interface-meta==1.3.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
joypy==0.2.6
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lightgbm==4.3.0
linearmodels==5.3
llvmlite==0.42.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.6
mca==1.0.3
mpl-finance==0.10.1
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.8.0
numba==0.59.1
numpy==1.26.2
openfe==0.0.12
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
patsy==0.5.3
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
pre-commit==3.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
Pygments==2.17.2
pyhdfe==0.2.0
pyparsing==3.1.1
pyproj==3.6.1
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
s3transfer==0.10.1
scikit-learn==1.3.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
setuptools-scm==7.1.0
shap==0.36.0
shapely==2.0.2
shiboken6==6.6.0
six==1.16.0
slicer==0.0.7
stack-data==0.6.3
statsmodels==0.14.0
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.4
traitlets==5.14.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.1
virtualenv==20.24.7
wcwidth==0.2.13
wrapt==1.16.0
xlrd==2.0.1
xyzservices==2023.10.1
